- Alterada a utilização da api do redmine da versão 1.17 para versão 2.0.0; - Alterado para retornar as versões com status aberto na combo da transação de geração dos cartões; - Fix: Na listagem dos cartões mostrar o total de pontos da sprint ao lado do botão de Exportar;

+++ src/main/java/br/com/voiza/rse/mbean/StoryCardMBean.java
@@ -71,7 +71,7 @@ public class StoryCardMBean extends SuperMBean implements Serializable {
     
     private Integer versionId;
     
-    private Integer totalPoints;
+    private double totalPoints;
     
     private boolean renderListagem;
     
@@ -204,11 +204,11 @@ public class StoryCardMBean extends SuperMBean implements Serializable {
         this.selectedIssues = selectedIssues;
     }
 
-    public Integer getTotalPoints() {
+    public double getTotalPoints() {
         return totalPoints;
     }
 
-    public void setTotalPoints(Integer totalPoints) {
+    public void setTotalPoints(double totalPoints) {
         this.totalPoints = totalPoints;
     }
